+++ tests/test_discover.py
@@ -24,49 +24,90 @@ def test_do_not_fail_when_extension_mod_has_not_all():
 def test_discover_all_activities():
     mod = "fixtures.fakeext"
     activities = discover_activities(mod, "probe")
-    assert len(activities) == 7
+    assert len(activities) == 8
 
-    activity = activities[0]
+    activities = iter(activities)
+
+    activity = next(activities)
     assert activity["name"] == "many_args"
     assert activity["type"] == "probe"
     assert activity["mod"] == mod
     assert activity["doc"] == "Many arguments."
     assert activity["arguments"] == [
         {
-            "name": "message"
+            "name": "message",
+            "type": "string"
         },
         {
             "name": "colour",
-            "default": "blue"
+            "default": "blue",
+            "type": "string"
         }
     ]
 
-    activity = activities[1]
+    activity = next(activities)
+    assert activity["name"] == "many_args_with_rich_types"
+    assert activity["type"] == "probe"
+    assert activity["mod"] == mod
+    assert activity["doc"] == "Many arguments with rich types."
+    assert activity["arguments"] == [
+        {
+            "name": "message",
+            "type": "string"
+        },
+        {
+            "name": "recipients",
+            "type": "list"
+        },
+        {
+            "name": "colour",
+            "default": "blue",
+            "type": "string"
+        },
+        {
+            "name": "count",
+            "default": 1,
+            "type": "integer"
+        },
+        {
+            "name": "logit",
+            "default": False,
+            "type": "boolean"
+        },
+        {
+            "name": "other",
+            "default": None,
+            "type": "object"
+        }
+    ]
+
+    activity = next(activities)
     assert activity["name"] == "no_args"
     assert activity["type"] == "probe"
     assert activity["mod"] == mod
     assert activity["doc"] == "No arguments."
     assert activity["arguments"] == []
 
-    activity = activities[2]
+    activity = next(activities)
     assert activity["name"] == "no_args_docstring"
     assert activity["type"] == "probe"
     assert activity["mod"] == mod
     assert activity["doc"] == None
     assert activity["arguments"] == []
 
-    activity = activities[3]
+    activity = next(activities)
     assert activity["name"] == "one_arg"
     assert activity["type"] == "probe"
     assert activity["mod"] == mod
     assert activity["doc"] == "One typed argument."
     assert activity["arguments"] == [
         {
-            "name": "message"
+            "name": "message",
+            "type": "string"
         }
     ]
 
-    activity = activities[4]
+    activity = next(activities)
     assert activity["name"] == "one_arg_with_default"
     assert activity["type"] == "probe"
     assert activity["mod"] == mod
@@ -74,11 +115,12 @@ def test_discover_all_activities():
     assert activity["arguments"] == [
         {
             "name": "message",
-            "default": "hello"
+            "default": "hello",
+            "type": "string"
         }
     ]
 
-    activity = activities[5]
+    activity = next(activities)
     assert activity["name"] == "one_untyped_arg"
     assert activity["type"] == "probe"
     assert activity["mod"] == mod
@@ -89,7 +131,7 @@ def test_discover_all_activities():
         }
     ]
 
-    activity = activities[6]
+    activity = next(activities)
     assert activity["name"] == "one_untyped_arg_with_default"
     assert activity["type"] == "probe"
     assert activity["mod"] == mod